requirements.txt
absl-py==0.7.1
astor==0.7.1
backcall==0.1.0
blinker==1.3
certifi==2019.3.9
chardet==3.0.4
cloud-init==17.2
command-not-found==0.3
configobj==5.0.6
cryptography==1.2.3
decorator==4.4.0
Django==2.1.7
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
idna==2.8
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
Jinja2==2.8
jsonpatch==1.10
jsonpointer==1.9
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
language-selector==0.1
Markdown==3.1
MarkupSafe==0.23
mock==2.0.0
numpy==1.16.2
oauthlib==1.0.3
parso==0.3.4
pbr==5.1.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prettytable==0.7.2
prompt-toolkit==2.0.9
protobuf==3.7.1
ptyprocess==0.6.0
pyasn1==0.1.9
pycurl==7.43.0
Pygments==2.3.1
pygobject==3.20.0
PyJWT==1.3.0
pyserial==3.0.1
python-apt==1.1.0b1+ubuntu0.16.4.1
python-debian==0.1.27
python-systemd==231
pytz==2017.2
PyYAML==5.1
requests==2.21.0
scipy==1.2.1
six==1.12.0
ssh-import-id==5.5
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
traitlets==4.3.2
ufw==0.35
unattended-upgrades==0.1
urllib3==1.24.1
virtualenv==16.4.0
wcwidth==0.1.7
Werkzeug==0.15.1
